+++ config/task/task_test.go
@@ -1,6 +1,7 @@
 package task
 
 import (
+	"os"
 	"testing"
 
 	"github.com/rliebz/tusk/config/marshal"
@@ -131,4 +132,62 @@ func TestTask_run_sub_tasks(t *testing.T) {
 	if err := task.run(r); err == nil {
 		t.Error(`task.run([exit 0, exit 1]): expected error, got nil`)
 	}
+}
+
+func TestTask_run_environment(t *testing.T) {
+	toBeUnset := "TO_BE_UNSET"
+	toBeUnsetValue := "unsetvalue"
+
+	toBeSet := "TO_BE_SET"
+	toBeSetValue := "setvalue"
+
+	if err := os.Setenv(toBeUnset, toBeUnsetValue); err != nil {
+		t.Fatalf(
+			"os.Setenv(%s, %s): unexpected error: %v",
+			toBeUnset, toBeUnsetValue, err,
+		)
+	}
+
+	defer func() {
+		if err := os.Unsetenv(toBeSet); err != nil {
+			t.Errorf(
+				"os.Unsetenv(%s): unexpected error: %v",
+				toBeSet, err,
+			)
+		}
+		if err := os.Unsetenv(toBeUnset); err != nil {
+			t.Errorf(
+				"os.Unsetenv(%s): unexpected error: %v",
+				toBeUnset, err,
+			)
+		}
+	}()
+
+	var task Task
+
+	r := &run.Run{
+		Environment: map[string]*string{
+			toBeSet:   &toBeSetValue,
+			toBeUnset: nil,
+		},
+	}
+
+	if err := task.run(r); err != nil {
+		t.Errorf("task.run(): unexpected error: %s", err)
+	}
+
+	if actual := os.Getenv(toBeSet); toBeSetValue != actual {
+		t.Errorf(
+			`value for %s: expected: "%s", actual: "%s"`,
+			toBeSet, toBeSetValue, actual,
+		)
+	}
+
+	if actual, isSet := os.LookupEnv(toBeUnset); isSet {
+		t.Errorf(
+			`value for %s: expected env var to be unset, actual: %s`,
+			toBeUnset, actual,
+		)
+	}
+
 }