+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Environment variables can now be set and unset inside of a `run` clause. This
+  replaces the `export` functionality that was previously under `option`.
+
 ### Changed
 - Log-level messages (Debug, Info, Warning, and Error) are now printed in title
   case instead of all caps.
+
+### Deprecated
+- Environment variables should no longer be set inside of an `option` clause.
 
 
 ## 0.2.0 (2017-11-08)